requirements.txt
--index-url https://pypi.python.org/simple/

beautifulsoup4==4.6.3
docopt==0.6.2
pyjq==2.1.0
PyYAML==3.13
requests==2.12.5
urllib3==1.23
